pyproject.toml
[tool.poetry]
name = "subspace"
version = "0.1.0"
description = ""
authors = ["Nikita Dorozhko <nick.dorozhko@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^6.0.4"
psycopg2 = "^2.9.12"
pillow = "^12.2.0"
djangorestframework = "^3.17.1"
djangorestframework-simplejwt = "^5.5.1"
python-dotenv = "^1.2.2"
stripe = "^15.1.0"
celery = "^5.6.3"
redis = "^7.4.0"
django-celery-beat = "^2.9.0"
dj-database-url = "^3.1.2"
django-cors-headers = "^4.9.0"
drf-yasg = "^1.21.15"
gunicorn = "^26.0.0"


[tool.poetry.group.dev.dependencies]
black = "^26.3.1"
flake8 = "^7.3.0"
isort = "^8.0.1"
pytest = "^9.0.3"
factory-boy = "^3.3.3"
coverage = "^7.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
